Commonize PointerMoveFilter
