+++ compose/ui/ui/src/commonMain/kotlin/androidx/compose/ui/input/pointer/PointerMoveFilter.kt
@@ -16,9 +16,13 @@
 
 package androidx.compose.ui.input.pointer
 
-import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.unit.IntSize
+
+// TODO: this code is copy-pasted from desktop (consider reusing it)
 
 /**
  * Modifier allowing to track pointer (i.e. mouse or trackpad) move events.
@@ -27,30 +31,32 @@ import androidx.compose.ui.geometry.Offset
  *  @param onEnter The callback invoked when pointer enters the component
  *  @param onExit The callback invoked when pointer leaves the component
  */
-@ExperimentalComposeUiApi
 fun Modifier.pointerMoveFilter(
     onMove: (position: Offset) -> Boolean = { false },
     onExit: () -> Boolean = { false },
     onEnter: () -> Boolean = { false },
-): Modifier = pointerInput(onMove, onExit, onEnter) {
-    awaitPointerEventScope {
-        while (true) {
-            val event = awaitPointerEvent()
-            val consumed = when (event.type) {
-                PointerEventType.Move -> {
-                    onMove(event.changes.first().position)
-                }
-                PointerEventType.Enter -> {
-                    onEnter()
-                }
-                PointerEventType.Exit -> {
-                    onExit()
-                }
-                else -> false
-            }
-            if (consumed) {
-                event.changes.forEach { it.consumeAllChanges() }
-            }
-        }
-    }
+): Modifier = composed {
+    val filter = remember(::PointerMoveEventFilter)
+    filter.onEnterHandler = onEnter
+    filter.onExitHandler = onExit
+    filter.onMoveHandler = onMove
+    MovePointerInputModifierImpl(filter)
 }
+
+internal class PointerMoveEventFilter : PointerInputFilter() {
+    lateinit var onEnterHandler: () -> Boolean
+    lateinit var onExitHandler: () -> Boolean
+    lateinit var onMoveHandler: (position: Offset) -> Boolean
+
+    override fun onPointerEvent(
+        pointerEvent: PointerEvent,
+        pass: PointerEventPass,
+        bounds: IntSize
+    ) = Unit
+
+    override fun onCancel() = Unit
+}
+
+private data class MovePointerInputModifierImpl(
+    override val pointerInputFilter: PointerInputFilter
+) : PointerInputModifier
